app/requirements.txt
fastapi
uvicorn[standard]
boto3
requests
cryptography
python-multipart
jinja2
reedsolo
sqlalchemy
PyJWT
passlib[bcrypt]
